// src/app/api/generated/model/adjunto-page-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdjuntoResponse } from './adjunto-response';


/**
 * Pagina de adjuntos administrativos de una persona
 */
export interface AdjuntoPageResponse { 
    /**
     * Adjuntos de esta pagina, mas nuevos primero
     */
    content?: Array<AdjuntoResponse>;
    /**
     * Numero de pagina devuelta, base cero
     */
    page?: number;
    /**
     * Cantidad de elementos por pagina
     */
    size?: number;
    /**
     * Cantidad total de adjuntos que cumplen el filtro
     */
    totalElements?: number;
    /**
     * Cantidad total de paginas disponibles
     */
    totalPages?: number;
}